requirements.txt
paho-mqtt>=2.1,<3
flask>=3.0,<4
waitress>=3.0,<4
requests>=2.31,<3
bcrypt>=4.1,<5
